서버 프로젝트 아키텍처 문서 작성
