--- app-server/subprojects/bounded_context/place/application/build.gradle.kts
@@ -1,3 +1,0 @@
-dependencies {
-    implementation(projects.domainEvent)
-}
