+++ app-server/subprojects/bounded_context/place/infra/build.gradle.kts
@@ -9,10 +9,12 @@ dependencies {
 
     implementation(projects.stdlib)
     implementation(projects.persistenceModel)
+    implementation(projects.domainEvent)
 
     implementation("org.springframework.boot:spring-boot-starter-web")
     implementation("org.springframework:spring-webflux")
     implementation("io.projectreactor.netty:reactor-netty")
+
 
     implementation("org.jetbrains.kotlinx:kotlinx-coroutines-reactive")
     implementation("org.jetbrains.kotlinx:kotlinx-serialization-json:$kotlinxSerializationVersion")
